+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 Changelog
 =========
+## 0.3.0 [UNRELEASED]
+### Bugfix
+  - CoNLL-U parsing now support comment lines, empty nodes and multiword tokens
+
+### API changes
+  - libginger
+    - CoNLL-U parsing now raise a specific `ParsingError` exception on parsing errors with a more
+      informative message
+    - Renamed `Tree.subtree()` → `Tree.descendance()` in order to allow adding a `Tree.subtree()`
+      function that will really return a `Tree` instead of a list of nodes.
 
 ## 0.2.1 [2017-03-05]
 ### Improvements